+++ aliyun-java-sdk-dataworks-public/src/main/java/com/aliyuncs/dataworks_public/model/v20200518/ListQualityRulesResponse.java
@@ -143,7 +143,7 @@ public class ListQualityRulesResponse extends AcsResponse {
 
 			private String comment;
 
-			private Integer ruleCheckerRelationId;
+			private Long ruleCheckerRelationId;
 
 			private Boolean fixCheck;
 
@@ -171,13 +171,13 @@ public class ListQualityRulesResponse extends AcsResponse {
 
 			private Integer checkerId;
 
-			private Integer entityId;
+			private Long entityId;
 
 			private String expectValue;
 
 			private Integer templateId;
 
-			private Integer id;
+			private Long id;
 
 			private String ruleName;
 
@@ -237,11 +237,11 @@ public class ListQualityRulesResponse extends AcsResponse {
 				this.comment = comment;
 			}
 
-			public Integer getRuleCheckerRelationId() {
+			public Long getRuleCheckerRelationId() {
 				return this.ruleCheckerRelationId;
 			}
 
-			public void setRuleCheckerRelationId(Integer ruleCheckerRelationId) {
+			public void setRuleCheckerRelationId(Long ruleCheckerRelationId) {
 				this.ruleCheckerRelationId = ruleCheckerRelationId;
 			}
 
@@ -349,11 +349,11 @@ public class ListQualityRulesResponse extends AcsResponse {
 				this.checkerId = checkerId;
 			}
 
-			public Integer getEntityId() {
+			public Long getEntityId() {
 				return this.entityId;
 			}
 
-			public void setEntityId(Integer entityId) {
+			public void setEntityId(Long entityId) {
 				this.entityId = entityId;
 			}
 
@@ -373,11 +373,11 @@ public class ListQualityRulesResponse extends AcsResponse {
 				this.templateId = templateId;
 			}
 
-			public Integer getId() {
+			public Long getId() {
 				return this.id;
 			}
 
-			public void setId(Integer id) {
+			public void setId(Long id) {
 				this.id = id;
 			}
